+++ src/commands/lane.ts
@@ -2,8 +2,8 @@
 import { Contract, ZeroAddress } from 'ethers'
 import type { TypedContract } from 'ethers-abitype'
 
-import FeeQuoterABI from '../abi/FeeQuoter_1_6.js'
-import RouterABI from '../abi/Router.js'
+import FeeQuoterABI from '../abi/FeeQuoter_1_6.ts'
+import RouterABI from '../abi/Router.ts'
 import {
   type CCIPContract,
   type CCIPVersion,
@@ -19,15 +19,10 @@ import {
   getTypeAndVersion,
   toObject,
   validateContractType,
-} from '../lib/index.js'
-import type { Providers } from '../providers.js'
-import { formatDuration, prettyLane } from './utils.js'
-
-export enum Format {
-  log = 'log',
-  pretty = 'pretty',
-  json = 'json',
-}
+} from '../lib/index.ts'
+import type { Providers } from '../providers.ts'
+import { Format } from './types.ts'
+import { formatDuration, prettyLane } from './utils.ts'
 
 export async function showLaneConfigs(
   providers: Providers,