+++ src/commands/show.ts
@@ -10,16 +10,16 @@ import {
   fetchCCIPMessagesInTx,
   fetchCommitReport,
   fetchExecutionReceipts,
-} from '../lib/index.js'
-import type { Providers } from '../providers.js'
-import { Format } from './types.js'
+} from '../lib/index.ts'
+import type { Providers } from '../providers.ts'
+import { Format } from './types.ts'
 import {
   prettyCommit,
   prettyReceipt,
   prettyRequest,
   selectRequest,
   withDateTimestamp,
-} from './utils.js'
+} from './utils.ts'
 
 export async function showRequests(
   providers: Providers,