+++ src/commands/utils.ts
@@ -20,10 +20,10 @@ import {
 } from 'ethers'
 import type { TypedContract } from 'ethers-abitype'
 
-import TokenABI from '../abi/BurnMintERC677Token.js'
-import TokenPoolABI from '../abi/BurnMintTokenPool_1_5_1.js'
-import RouterABI from '../abi/Router.js'
-import TokenAdminRegistry from '../abi/TokenAdminRegistry_1_5.js'
+import TokenABI from '../abi/BurnMintERC677Token.ts'
+import TokenPoolABI from '../abi/BurnMintTokenPool_1_5_1.ts'
+import RouterABI from '../abi/Router.ts'
+import TokenAdminRegistry from '../abi/TokenAdminRegistry_1_5.ts'
 import {
   type CCIPCommit,
   type CCIPContract,
@@ -43,11 +43,14 @@ import {
   networkInfo,
   parseWithFragment,
   recursiveParseError,
-} from '../lib/index.js'
+} from '../lib/index.ts'
 
 export async function getWallet(argv?: { wallet?: string }): Promise<Signer> {
   if ((argv?.wallet ?? '').startsWith('ledger')) {
-    const ledger = await LedgerSigner.create(null, argv!.wallet!.split(':')[1])
+    let derivationPath = argv!.wallet!.split(':')[1]
+    if (derivationPath && !isNaN(Number(derivationPath)))
+      derivationPath = `m/44'/60'/${derivationPath}'/0/0`
+    const ledger = await LedgerSigner.create(null, derivationPath)
     console.info('Ledger connected:', await ledger.getAddress(), `at "${ledger.path}"`)
     return ledger
   }