+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- `--wallet ledger:<n>` supports an integer as wallet index on standard Ethereum derivation path (#24)
 
 ## [0.2.1] - 2025-03-11
 - Fix manualExec on old v1.2 lanes or <1.5 TokenPools (legacy sourceTokenData) (#18)